+++ meshlink-reference/ios/ReferenceApp/ReferenceApp.swift
@@ -1,4 +1,4 @@
-import ReferenceAppShared
+import MeshLinkReference
 import SwiftUI
 import UIKit